scene refactor
